+++ changelog.md
@@ -37,7 +37,7 @@
 
 - Removed deprecated `iup` module from stdlib, it has already moved to [nimble](https://github.com/nim-lang/iup).
 
-- `doAssertRaises` now correctly handles foreign exceptions; it also allows a catch-all form that includes foreign exceptions.
+- `doAssertRaises` now correctly handles foreign exceptions.
 
 ## Language changes